Install YARN on cgcloud-jenkins/toil_jenkins_slave (resolves #149)
Installs Apache Hadoop 2.6.2 on the toil_jenkins_slave image by pulling down
the prebuilt artifacts from an Apache mirror. YARN is a subcomponent of Hadoop.
This does not start the YARN nodemanager/resourcemanager, as they are started
downstream in the toil test suites.

Also see https://github.com/BD2KGenomics/toil/issues/796.
